Testing binDensity for extensive fields

+++ binDensity.py
@@ -703,7 +703,7 @@ def densityBin(fileT,fileS,fileFx,outFile,debug=True,timeint='all',mthout=False)
             ssr[0,:] = ssr[1,:]
             #print 's_s,ssr',s_s[:,ijtest],ssr[:,ijtest]
             inds_bottom = npy.argwhere ( (szmax <= s_s) & (szmax > ssr) ).transpose()
-            bottom_ind = npy.zeros(lonN*latN)
+            bottom_ind = npy.zeros(lonN*latN, dtype='int8')
             print bottom_ind.shape
             print  inds_bottom.shape, inds.shape
             print s_s[inds[0][npy.argwhere (inds[1] == ijtest)],ijtest]
